split mcap coverage by behavior
